search-blitz: move loadQueries into config.go
This logic seems like it should live with the config structures.
